requirements.txt
APScheduler==3.10.3
asgiref==3.7.2
boto3==1.28.40
botocore==1.31.40
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==3.1.1
django-admin-volt==1.0.10
django-apscheduler==0.6.2
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
django-sotp==1.0.4
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.0.5
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
pymemcache==4.0.0
pyotp==2.9.0
python-dotenv==0.9.1
python-environ==0.4.54
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rest-api-payload==0.0.7
rest-swagger==1.0.0
s3transfer==0.6.2
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==6.5.0
